C++ implementation of the file parser
I also uploaded a new dictionary file. This is by far not the full
implementation of the file parser. All it does is asks the user to
specify the number of characters they want to use and after that if
shows the random letters using a function I made and then it displays
all words in the file from 2 to (n) characters.